% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Scan specified columns for unique responses.}
\usage{
assess(df, cols)
}
\arguments{
\item{df}{A dataframe or tibble.}

\item{cols}{A column index or vector of column indices.}
}
\description{
Scan specified columns for unique responses.
}
