% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify.R
\name{unify}
\alias{unify}
\title{Aggregate survey data into a tidy format.}
\usage{
unify(
  df,
  cols,
  ...,
  ignore = NULL,
  filter = NULL,
  tibble = TRUE,
  gtTable = FALSE,
  colSplit = FALSE,
  hideResponse = FALSE,
  hideN = FALSE,
  hideProportion = FALSE
)
}
\arguments{
\item{df}{A dataframe or tibble.}

\item{cols}{A column index or vector of column indices.}

\item{...}{Groups of character vectors containing responses to be aggregated together. Each group is a named list of responses.}

\item{ignore}{A character vector of responses to be ignored from the table and calculations.}
}
\value{
An aggregated table of responses within specified groups.
}
\description{
Aggregate survey data into a tidy format.
}
