% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify.R
\name{unify}
\alias{unify}
\title{Aggregate survey data into a tidy format.}
\usage{
unify(
  df,
  cols,
  ...,
  ignore = NULL,
  filter = NULL,
  gtTable = FALSE,
  colSplit = FALSE,
  hideResponse = FALSE,
  hideN = FALSE,
  hideProportion = FALSE
)
}
\arguments{
\item{df}{A dataframe or tibble.}

\item{cols}{A column index or vector of column indices.}

\item{...}{Groups containing a string or string vector conforming to responses in the specified column(s).}

\item{ignore}{A string or string vector of responses to be ignored from the table and calculations.}

\item{filter}{A string or string vector of responses to filter for the output.}

\item{gtTable}{A logical value indicating whether to output as a gt table.}

\item{colSplit}{A logical value indicating whether to split responses into separate columns in the output.}

\item{hideResponse}{A logical value indicating whether to hide the response field from the output.}

\item{hideN}{A logical value indicating whether to hide the n field from the output.}

\item{hideProportion}{A logical value indicating whether to hide the proportion field from the output.}
}
\value{
An aggregated table of responses within specified groups.
}
\description{
Aggregate survey data into a tidy format.
}
